Move files and adjust references
