+++ app/src/main/java/com/mypackage/sportsevents/presentation/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.mypackage.sportsevents
+package com.mypackage.sportsevents.presentation.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.mypackage.sportsevents.ui.theme.SportsEventsTheme
+import com.mypackage.sportsevents.presentation.ui.theme.SportsEventsTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
