+++ app/src/main/java/com/mypackage/sportsevents/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.mypackage.sportsevents.ui.theme
+package com.mypackage.sportsevents.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
